Add metadata and styles to posts
Metadata is added as frontmatter variables. Jekyll supports these out
of the box and allows to configure jekyl and our langing page neatly in
one place.

During CI the inject script parses the metadata comments from posts
to inject the properties to the posts on the timeline. This includes
sorting them in a descending order using the date.

Add styles for tags to wrap around when there are many tags.